refactor: Separated Discord Handling Logic from handlers.go
